midx.c: unify `include_pack` array into `pack_info` struct
The MIDX repack code uses a separate byte-array to track which packs are
"included" in the rollup. Now that an array of `repack_info` structs is
available for the lifetime of the `midx_repack()` function, we can
instead use a new bit in that struct's record instead of maintaining a
separate array.

Signed-off-by: Taylor Blau <me@ttaylorr.com>