midx.c: compute pack_info array outside of fill_included_packs_batch()
The fill_included_packs_batch() function needs to compute a list of
packs which are included in the MIDX, and then sort that list according
to the pack mtimes.

Since packs are listed in the `m->pack_names` array in SHA-1 order, we
have to reconstruct their mtimes by reading the `pack_mtime` member of
the struct `packed_git` in order to determine the value by which to
sort.

This list is only needed by the fill_included_packs_batch(), but the
subsequent commit will want to have this list accessible in this
function's caller, `midx_repack()`.

Prepare for that by teaching `midx_repack()` to compute, pass in, and
then free that list so it is available throughout the body of
`midx_repack()`.

Signed-off-by: Taylor Blau <me@ttaylorr.com>