config: relax requirements on multi-value return
The git_configset_get_value() method has an assert() statement
guaranteeing that the result from git_configset_get_value_multi() is
either NULL or has at least one element. We want to change that return
to provide an empty list instead of a NULL list, so change the earlier
'return 1' condition to care about a NULL or empty list.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>