bundle: add flags to verify_bundle(), skip walk
The verify_bundle() method checks if a bundle can be applied to a given
repository. This not only verifies that certain commits exist in the
repository, but Git also checks that these commits are reachable.

This behavior dates back to the original git-bundle builtin written in
2e0afafebd8 (Add git-bundle: move objects and references by archive,
2007-02-22), but the message does not go into detail why the
reachability check is important.

Since verify_bundle() is called from unbundle(), we need to add an
option to pipe the flags through that method.

When unbundling from a list of bundles, Git will create refs that point
to the tips of the latest bundle, which makes this reachability walk
succeed, in theory. However, the loose refs cache does not get
invalidated and hence the reachability walk fails. By disabling the
reachability walk in the bundle URI code, we can get around this
reachability check.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>