documentation: format-patch: clarify requirements for patch-ids to match
The documentation for format-patch advertises that the ids
it generates for prerequisite patches are the same as piping
the patch into "git patch-id". Clarify here that this is only
true if the patch was generated with -U3, and for binary patches,
with --full-index.

Note that the actual equivalence isn't currently tested. Aside
from a few cases fixed in this patch series, I've seen some
uncommon situations where "git diff" and the internal diff api
actually generate equally valid diffs of the same length, but
with lines reordered, which results in different patch-ids.

Signed-off-by: Jerry Zhang <Jerry@skydio.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>