format-patch: introduce branch-scoped variables
Introduce variants of many of the `format.*` configuration variables
with `format.<branch>.*` counterparts that only take effect on a
matching branch, if applicable.

This can be useful when shuffling multiple topics around, where each of
the topics has a different set of recipients.

This topic makes it possible to permanently add, for example, 'To:' or
'Cc:' recipients specific to each branch, such that running
`format-patch` on that branch automatically includes the appropriate
recipients.

Signed-off-by: Taylor Blau <me@ttaylorr.com>