Meta/Reintegrate: attach a tty during merge annotation
Reintegrate assumes that your $GIT_EDITOR does not require a tty in
order to annotate any merges during reintegration. But some editors,
like Vi, are useless without a tty.

Make sure that we redirect /dev/tty over stdin to our editor to make
sure that it works with Vi/Vim.

Signed-off-by: Taylor Blau <me@ttaylorr.com>